requirements.txt
mysql-connector-python>=8.0.29
botocore>=1.27.23
boto3>=1.24.23
fastapi>=0.100.0